+++ HISTORY.md
@@ -54,6 +54,7 @@
 
 ### 2022-06-25
 * threads: made slot-type mnemonics case-insensitive
+* threads: made categories case-insensitive
 
 ### 2022-06-24
 * threads: made block attribute "type" case-insensitive for textual mnemonics